organizando os controladores
