+++ src/main/java/com/mycompany/darktools/controller/viwer/SaveController.java
@@ -3,22 +3,36 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.mycompany.darktools.viwer;
+package com.mycompany.darktools.controller.viwer;
 
 import java.net.URL;
 import java.util.ResourceBundle;
+import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.scene.control.Button;
+import javafx.scene.control.CheckBox;
 
 /**
  * FXML Controller class
  *
  * @author Rafae
  */
-public class CustomPersonController implements Initializable {
+public class SaveController implements Initializable {
 
-    /**
-     * Initializes the controller class.
-     */
+    
+    
+    @FXML
+    private CheckBox cb_Easy;
+
+    @FXML
+    private CheckBox cb_medium;
+
+    @FXML
+    private CheckBox cb_hard;
+
+    @FXML
+    private Button btn_exit;
+    
     @Override
     public void initialize(URL url, ResourceBundle rb) {
         // TODO
